VSB-TUO/Test: share the embargo integration test fixture
The six ItemUpdate embargo test classes each carried their own copy of the same
setup: the collection whose bitstreams inherit an undated Anonymous READ policy,
the metadata field registration, the SAF archive writer that drives a run, and
the policy inspection helpers.

AbstractEmbargoIT now holds that fixture and each class keeps only the scenario
it pins down. No test case is removed and no assertion changes.

Three helpers had drifted apart between the copies and are unified on the
version that was already documented as the correct one:

- replaceAnonymousReadPolicies deletes policies one by one; the bulk
  removePoliciesActionFilter leaves the in-memory collection stale
- dublinCore has one implementation instead of five
- a run tees the console, so the output still reaches the failsafe report
