+++ dspace-api/src/test/java/org/dspace/app/itemupdate/EmbargoDerivativesIT.java
@@ -12,56 +12,18 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.PrintStream;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.sql.SQLException;
 import java.time.LocalDate;
-import java.time.ZoneId;
-import java.time.ZoneOffset;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
-import java.util.Date;
 import java.util.List;
 import java.util.Set;
-import java.util.TreeSet;
-import java.util.stream.Collectors;
 
-import org.apache.commons.io.file.PathUtils;
-import org.dspace.AbstractIntegrationTestWithDatabase;
 import org.dspace.authorize.ResourcePolicy;
-import org.dspace.authorize.factory.AuthorizeServiceFactory;
-import org.dspace.authorize.service.AuthorizeService;
-import org.dspace.authorize.service.ResourcePolicyService;
-import org.dspace.builder.BitstreamBuilder;
-import org.dspace.builder.CollectionBuilder;
-import org.dspace.builder.CommunityBuilder;
-import org.dspace.builder.ItemBuilder;
-import org.dspace.builder.MetadataFieldBuilder;
-import org.dspace.builder.ResourcePolicyBuilder;
 import org.dspace.content.Bitstream;
 import org.dspace.content.Bundle;
-import org.dspace.content.Collection;
 import org.dspace.content.Item;
-import org.dspace.content.MetadataField;
-import org.dspace.content.MetadataSchema;
-import org.dspace.content.factory.ContentServiceFactory;
-import org.dspace.content.service.ItemService;
-import org.dspace.content.service.MetadataFieldService;
-import org.dspace.content.service.MetadataSchemaService;
 import org.dspace.core.Constants;
-import org.dspace.eperson.EPerson;
-import org.dspace.eperson.Group;
-import org.dspace.eperson.factory.EPersonServiceFactory;
-import org.dspace.eperson.service.GroupService;
-import org.dspace.handle.factory.HandleServiceFactory;
-import org.dspace.handle.service.HandleService;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 
 /**
@@ -70,72 +32,14 @@ import org.junit.Test;
  * re-opened when the embargo ends. Bundles that are not derived from the file, LICENSE above all, stay out
  * of scope.
  */
-public class EmbargoDerivativesIT extends AbstractIntegrationTestWithDatabase {
-
-    /** Policy name the synchronisation writes. */
-    private static final String EMBARGO_POLICY_NAME = "embargo";
+public class EmbargoDerivativesIT extends AbstractEmbargoIT {
 
     private static final String OPEN_ACCESS = "openAccess";
     private static final String EMBARGOED_ACCESS = "embargoedAccess";
     private static final String RESTRICTED_ACCESS = "restrictedAccess";
 
-    private static final String TEXT_BUNDLE = "TEXT";
-    private static final String THUMBNAIL_BUNDLE = "THUMBNAIL";
     private static final String LICENSE_BUNDLE = "LICENSE";
     private static final String CC_LICENSE_BUNDLE = "CC-LICENSE";
-
-    private final ItemService itemService = ContentServiceFactory.getInstance().getItemService();
-    private final HandleService handleService = HandleServiceFactory.getInstance().getHandleService();
-    private final ResourcePolicyService resourcePolicyService =
-            AuthorizeServiceFactory.getInstance().getResourcePolicyService();
-    private final AuthorizeService authorizeService = AuthorizeServiceFactory.getInstance().getAuthorizeService();
-    private final GroupService groupService = EPersonServiceFactory.getInstance().getGroupService();
-    private final MetadataSchemaService metadataSchemaService =
-            ContentServiceFactory.getInstance().getMetadataSchemaService();
-    private final MetadataFieldService metadataFieldService =
-            ContentServiceFactory.getInstance().getMetadataFieldService();
-
-    private Collection collection;
-    private Group anonymousGroup;
-    private Path tempDir;
-    private String previousHandlePrefix;
-
-    @Before
-    @Override
-    public void setUp() throws Exception {
-        super.setUp();
-        context.turnOffAuthorisationSystem();
-
-        parentCommunity = CommunityBuilder.createCommunity(context)
-                .withName("Parent Community")
-                .build();
-        collection = CollectionBuilder.createCollection(context, parentCommunity)
-                .withName("Collection")
-                .build();
-
-        // neither field exists in the test metadata registry, AddMetadataAction needs both
-        ensureMetadataFieldExists("rights", "access");
-        ensureMetadataFieldExists("date", "embargoend");
-
-        anonymousGroup = groupService.findByName(context, Group.ANONYMOUS);
-        // ItemArchive resolves items through this mutable static, it is restored in destroy()
-        previousHandlePrefix = ItemUpdate.HANDLE_PREFIX;
-        ItemUpdate.HANDLE_PREFIX = handleService.getCanonicalPrefix();
-
-        context.restoreAuthSystemState();
-
-        tempDir = Files.createTempDirectory("embargoDerivativesIT");
-    }
-
-    @After
-    @Override
-    public void destroy() throws Exception {
-        ItemUpdate.HANDLE_PREFIX = previousHandlePrefix;
-        if (tempDir != null) {
-            PathUtils.deleteDirectory(tempDir);
-        }
-        super.destroy();
-    }
 
     /**
      * Verifies that a running embargo closes the thumbnail and the extracted full text as well. The TEXT
@@ -189,7 +93,8 @@ public class EmbargoDerivativesIT extends AbstractIntegrationTestWithDatabase {
 
         // the state a finished embargo run leaves behind: one dated Anonymous READ policy on every bundle
         for (Bitstream file : files) {
-            replaceAnonymousReadPolicy(file, startOfDayUtc(LocalDate.now().plusYears(1)), EMBARGO_POLICY_NAME);
+            replaceAnonymousReadPolicies(file, startOfDayUtc(LocalDate.now().plusYears(1)),
+                    EMBARGO_POLICY_NAME, ResourcePolicy.TYPE_CUSTOM);
         }
         reloadAll(files);
         for (Bitstream file : files) {
@@ -310,7 +215,7 @@ public class EmbargoDerivativesIT extends AbstractIntegrationTestWithDatabase {
         Bitstream text = createBitstreamInBundle(item, "closed.pdf.txt", TEXT_BUNDLE);
 
         context.turnOffAuthorisationSystem();
-        deleteReadPolicies(text);
+        deletePolicies(text, Constants.READ);
         context.restoreAuthSystemState();
         text = context.reloadEntity(text);
 
@@ -407,68 +312,6 @@ public class EmbargoDerivativesIT extends AbstractIntegrationTestWithDatabase {
     }
 
     /**
-     * Runs itemupdate with both embargo fields as targets, the combination that triggers embargo
-     * synchronisation.
-     *
-     * @return the console output of the run and the number of embargo problems it counted
-     */
-    private Run runItemUpdate(Item item, String dublinCoreContent) throws Exception {
-        Path sourceRoot = Files.createDirectory(tempDir.resolve("saf-" + System.nanoTime()));
-        // without this marker processArchive writes an undo archive next to the source directory
-        Files.createFile(sourceRoot.resolve(ItemUpdate.SUPPRESS_UNDO_FILENAME));
-
-        Path itemDir = Files.createDirectory(sourceRoot.resolve("item_000"));
-        Files.writeString(itemDir.resolve("dublin_core.xml"), dublinCoreContent, StandardCharsets.UTF_8);
-
-        ItemUpdate itemUpdate = new ItemUpdate();
-        DeleteMetadataAction deleteAction =
-                (DeleteMetadataAction) itemUpdate.actionMgr.getUpdateAction(DeleteMetadataAction.class);
-        deleteAction.addTargetFields(new String[] { "dc.rights.access", "dc.date.embargoend" });
-
-        AddMetadataAction addAction =
-                (AddMetadataAction) itemUpdate.actionMgr.getUpdateAction(AddMetadataAction.class);
-        addAction.addTargetFields(new String[] { "dc.rights.access", "dc.date.embargoend" });
-
-        // ItemUpdate reports to the console only, so it has to be captured to assert on it
-        ByteArrayOutputStream consoleBuffer = new ByteArrayOutputStream();
-        PrintStream originalOut = System.out;
-        PrintStream originalErr = System.err;
-        PrintStream captureStream = new PrintStream(consoleBuffer, true, StandardCharsets.UTF_8);
-
-        context.turnOffAuthorisationSystem();
-        System.setOut(captureStream);
-        System.setErr(captureStream);
-        try {
-            itemUpdate.processArchive(context, sourceRoot.toString(), null, null, true, false, true);
-        } finally {
-            captureStream.flush();
-            System.setOut(originalOut);
-            System.setErr(originalErr);
-            context.restoreAuthSystemState();
-        }
-
-        context.uncacheEntity(item);
-
-        String consoleOutput = consoleBuffer.toString(StandardCharsets.UTF_8);
-        // replay it so the failsafe -output.txt still holds the full ItemUpdate log
-        System.out.println(consoleOutput);
-        return new Run(consoleOutput, itemUpdate.embargoSyncFailures);
-    }
-
-    /**
-     * What a finished {@code itemupdate} run is judged by: its console output and its failure count.
-     */
-    private static final class Run {
-        private final String console;
-        private final int embargoSyncFailures;
-
-        private Run(String console, int embargoSyncFailures) {
-            this.console = console;
-            this.embargoSyncFailures = embargoSyncFailures;
-        }
-    }
-
-    /**
      * Identity and content of every policy of one bitstream, taken before a run.
      */
     private static final class Snapshot {
@@ -490,48 +333,6 @@ public class EmbargoDerivativesIT extends AbstractIntegrationTestWithDatabase {
     }
 
     /**
-     * Asserts the exit code the run would have produced; without it a skipped item looks synchronised.
-     */
-    private void assertExitCode(String scenario, int expectedFailures, Run run) {
-        assertEquals("[" + scenario + "] wrong number of reported embargo problems, so ItemUpdate.main() would"
-                        + " exit with " + ItemUpdate.exitStatus(0, run.embargoSyncFailures) + " instead of "
-                        + ItemUpdate.exitStatus(0, expectedFailures) + ". Console output was:"
-                        + System.lineSeparator() + run.console,
-                expectedFailures, run.embargoSyncFailures);
-    }
-
-    /**
-     * Builds a SAF {@code dublin_core.xml} carrying one {@code dc.rights.access} and one
-     * {@code dc.date.embargoend} value.
-     */
-    private String dublinCore(Item item, String accessRight, String embargoEndDate) {
-        return "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-                + "<dublin_core schema=\"dc\">\n"
-                + "    <dcvalue element=\"identifier\" qualifier=\"uri\">"
-                + ItemUpdate.HANDLE_PREFIX + item.getHandle() + "</dcvalue>\n"
-                + "    <dcvalue element=\"rights\" qualifier=\"access\">" + accessRight + "</dcvalue>\n"
-                + "    <dcvalue element=\"date\" qualifier=\"embargoend\">" + embargoEndDate + "</dcvalue>\n"
-                + "</dublin_core>";
-    }
-
-    private void ensureMetadataFieldExists(String element, String qualifier) throws Exception {
-        MetadataSchema dcSchema = metadataSchemaService.find(context, "dc");
-        MetadataField existingField = metadataFieldService.findByElement(context, dcSchema, element, qualifier);
-        if (existingField == null) {
-            MetadataFieldBuilder.createMetadataField(context, dcSchema, element, qualifier, null).build();
-        }
-    }
-
-    private Item createItem(String title) throws Exception {
-        context.turnOffAuthorisationSystem();
-        Item item = ItemBuilder.createItem(context, collection)
-                .withTitle(title)
-                .build();
-        context.restoreAuthSystemState();
-        return item;
-    }
-
-    /**
      * The state filter-media leaves behind on a public item: the file plus its extracted text and its
      * thumbnail, each with the one undated Anonymous READ policy inherited from the collection.
      */
@@ -543,91 +344,12 @@ public class EmbargoDerivativesIT extends AbstractIntegrationTestWithDatabase {
         return files;
     }
 
-    private Bitstream createBitstreamInBundle(Item item, String name, String bundleName) throws Exception {
-        context.turnOffAuthorisationSystem();
-        Bitstream bitstream = BitstreamBuilder.createBitstream(context, item,
-                        new ByteArrayInputStream(("content-" + name).getBytes(StandardCharsets.UTF_8)),
-                        bundleName)
-                .withName(name)
-                .withMimeType("text/plain")
-                .build();
-        context.restoreAuthSystemState();
-        return bitstream;
-    }
-
-    /**
-     * Replaces every READ policy of the bitstream with a single dated Anonymous READ policy.
-     */
-    private ResourcePolicy replaceAnonymousReadPolicy(Bitstream bitstream, Date startDate, String name)
-            throws Exception {
-        context.turnOffAuthorisationSystem();
-        deleteReadPolicies(bitstream);
-        ResourcePolicy policy = ResourcePolicyBuilder.createResourcePolicy(context, null, anonymousGroup)
-                .withAction(Constants.READ)
-                .withDspaceObject(bitstream)
-                .withName(name)
-                .withPolicyType(ResourcePolicy.TYPE_CUSTOM)
-                .withStartDate(startDate)
-                .build();
-        context.restoreAuthSystemState();
-        return policy;
-    }
-
-    /**
-     * Deletes READ policies one by one, because the bulk removal helpers issue an HQL delete and leave the
-     * in-memory collection stale.
-     */
-    private void deleteReadPolicies(Bitstream bitstream) throws Exception {
-        for (ResourcePolicy policy : readPolicies(bitstream)) {
-            resourcePolicyService.delete(context, policy);
-        }
-    }
-
-    private List<ResourcePolicy> readPolicies(Bitstream bitstream) throws SQLException {
-        return new ArrayList<>(resourcePolicyService.find(context, bitstream, Constants.READ));
-    }
-
-    private List<ResourcePolicy> anonymousReadPolicies(Bitstream bitstream) throws SQLException {
-        return readPolicies(bitstream).stream()
-                .filter(policy -> policy.getGroup() != null && anonymousGroup.equals(policy.getGroup()))
-                .collect(Collectors.toList());
-    }
-
     private ResourcePolicy onlyAnonymousReadPolicy(Bitstream bitstream, String what) throws SQLException {
         List<ResourcePolicy> policies = anonymousReadPolicies(bitstream);
         assertEquals("[" + label(bitstream) + "] exactly one Anonymous READ policy must remain after " + what
                         + ", a second one would defeat the embargo." + describe(bitstream),
                 1, policies.size());
         return policies.get(0);
-    }
-
-    private Set<Integer> policyIds(Bitstream bitstream) throws SQLException {
-        Set<Integer> ids = new TreeSet<>();
-        for (ResourcePolicy policy : resourcePolicyService.find(context, bitstream)) {
-            ids.add(policy.getID());
-        }
-        return ids;
-    }
-
-    private List<String> policyFingerprints(Bitstream bitstream) throws SQLException {
-        List<String> fingerprints = new ArrayList<>();
-        for (ResourcePolicy policy : resourcePolicyService.find(context, bitstream)) {
-            fingerprints.add(fingerprint(policy));
-        }
-        Collections.sort(fingerprints);
-        return fingerprints;
-    }
-
-    private String fingerprint(ResourcePolicy policy) {
-        return String.format("id=%s action=%s group=%s eperson=%s rpType=%s rpName=%s start=%s end=%s",
-                policy.getID(),
-                Constants.actionText[policy.getAction()],
-                policy.getGroup() == null ? "<none>" : policy.getGroup().getName(),
-                policy.getEPerson() == null ? "<none>" : policy.getEPerson().getEmail(),
-                policy.getRpType(),
-                policy.getRpName(),
-                day(policy.getStartDate()),
-                day(policy.getEndDate()));
     }
 
     /**
@@ -658,50 +380,4 @@ public class EmbargoDerivativesIT extends AbstractIntegrationTestWithDatabase {
         return (bundles.isEmpty() ? "<no bundle>" : bundles.get(0).getName()) + "/" + bitstream.getName();
     }
 
-    /**
-     * Tells whether a visitor who is not logged in may read the bitstream. The authorisation state is a stack
-     * the builders push and pop, so it is drained first - otherwise every read looks allowed.
-     */
-    private boolean anonymousCanRead(Bitstream bitstream) throws SQLException {
-        EPerson savedUser = context.getCurrentUser();
-        int popped = 0;
-        while (context.ignoreAuthorization()) {
-            context.restoreAuthSystemState();
-            popped++;
-        }
-        context.setCurrentUser(null);
-        try {
-            return authorizeService.authorizeActionBoolean(context, bitstream, Constants.READ);
-        } finally {
-            context.setCurrentUser(savedUser);
-            for (int i = 0; i < popped; i++) {
-                context.turnOffAuthorisationSystem();
-            }
-        }
-    }
-
-    private void reloadAll(List<Bitstream> bitstreams) throws SQLException {
-        for (int i = 0; i < bitstreams.size(); i++) {
-            bitstreams.set(i, context.reloadEntity(bitstreams.get(i)));
-        }
-    }
-
-    private Date startOfDayUtc(LocalDate day) {
-        return Date.from(day.atStartOfDay(ZoneOffset.UTC).toInstant());
-    }
-
-    /**
-     * Start dates come back from the database as {@code java.sql.Date}, so they are compared at calendar day
-     * granularity instead of as instants.
-     */
-    private LocalDate toLocalDate(Date date) {
-        if (date instanceof java.sql.Date) {
-            return ((java.sql.Date) date).toLocalDate();
-        }
-        return date.toInstant().atZone(ZoneId.systemDefault()).toLocalDate();
-    }
-
-    private String day(Date date) {
-        return date == null ? "<none>" : toLocalDate(date).toString();
-    }
 }
