+++ dspace-api/src/test/java/org/dspace/app/itemupdate/EmbargoSafetyIT.java
@@ -11,59 +11,24 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.PrintStream;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.sql.SQLException;
 import java.time.LocalDate;
-import java.time.ZoneId;
-import java.time.ZoneOffset;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.Date;
 import java.util.List;
 import java.util.Set;
-import java.util.TreeSet;
-import java.util.stream.Collectors;
 
-import org.apache.commons.io.file.PathUtils;
-import org.dspace.AbstractIntegrationTestWithDatabase;
 import org.dspace.authorize.AuthorizeException;
 import org.dspace.authorize.ResourcePolicy;
-import org.dspace.authorize.factory.AuthorizeServiceFactory;
-import org.dspace.authorize.service.AuthorizeService;
-import org.dspace.authorize.service.ResourcePolicyService;
-import org.dspace.builder.BitstreamBuilder;
-import org.dspace.builder.CollectionBuilder;
-import org.dspace.builder.CommunityBuilder;
 import org.dspace.builder.GroupBuilder;
-import org.dspace.builder.ItemBuilder;
-import org.dspace.builder.MetadataFieldBuilder;
 import org.dspace.builder.ResourcePolicyBuilder;
 import org.dspace.content.Bitstream;
-import org.dspace.content.Collection;
 import org.dspace.content.Item;
-import org.dspace.content.MetadataField;
-import org.dspace.content.MetadataSchema;
 import org.dspace.content.MetadataValue;
-import org.dspace.content.factory.ContentServiceFactory;
-import org.dspace.content.service.ItemService;
-import org.dspace.content.service.MetadataFieldService;
-import org.dspace.content.service.MetadataSchemaService;
 import org.dspace.core.Constants;
 import org.dspace.core.Context;
-import org.dspace.eperson.EPerson;
 import org.dspace.eperson.Group;
-import org.dspace.eperson.factory.EPersonServiceFactory;
-import org.dspace.eperson.service.GroupService;
-import org.dspace.handle.factory.HandleServiceFactory;
-import org.dspace.handle.service.HandleService;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 
 /**
@@ -71,13 +36,7 @@ import org.junit.Test;
  * rather than publish it. Every "must not touch" assertion compares policy ids and a fingerprint of every
  * policy, because counts hide both delete-and-recreate and in-place mutation.
  */
-public class EmbargoSafetyIT extends AbstractIntegrationTestWithDatabase {
-
-    /**
-     * rpName written by earlier versions; the fixtures use it so that normalisation of legacy policies is
-     * exercised.
-     */
-    private static final String LEGACY_EMBARGO_POLICY_NAME = "Standard Embargo";
+public class EmbargoSafetyIT extends AbstractEmbargoIT {
 
     /**
      * The supported way of re-opening files whose Anonymous READ policy is already gone; ItemUpdate points
@@ -90,65 +49,6 @@ public class EmbargoSafetyIT extends AbstractIntegrationTestWithDatabase {
      * again on that day.
      */
     private static final String LEASE_POLICY_NAME = "lease";
-
-    /**
-     * Sentinel for {@link #deletePolicies(Bitstream, int)} meaning "every action", picked so it can never
-     * collide with a real value of {@link Constants#actionText}.
-     */
-    private static final int ALL_ACTIONS = -1;
-
-    private final ItemService itemService = ContentServiceFactory.getInstance().getItemService();
-    private final HandleService handleService = HandleServiceFactory.getInstance().getHandleService();
-    private final ResourcePolicyService resourcePolicyService =
-            AuthorizeServiceFactory.getInstance().getResourcePolicyService();
-    private final AuthorizeService authorizeService = AuthorizeServiceFactory.getInstance().getAuthorizeService();
-    private final GroupService groupService = EPersonServiceFactory.getInstance().getGroupService();
-    private final MetadataSchemaService metadataSchemaService =
-            ContentServiceFactory.getInstance().getMetadataSchemaService();
-    private final MetadataFieldService metadataFieldService =
-            ContentServiceFactory.getInstance().getMetadataFieldService();
-
-    private Collection collection;
-    private Group anonymousGroup;
-    private Path tempDir;
-    private String previousHandlePrefix;
-
-    @Before
-    @Override
-    public void setUp() throws Exception {
-        super.setUp();
-        context.turnOffAuthorisationSystem();
-
-        parentCommunity = CommunityBuilder.createCommunity(context)
-                .withName("Parent Community")
-                .build();
-        collection = CollectionBuilder.createCollection(context, parentCommunity)
-                .withName("Collection")
-                .build();
-
-        // neither field exists in the test metadata registry, AddMetadataAction needs both
-        ensureMetadataFieldExists("rights", "access");
-        ensureMetadataFieldExists("date", "embargoend");
-
-        anonymousGroup = groupService.findByName(context, Group.ANONYMOUS);
-        // ItemArchive resolves items through this mutable static, it is restored in destroy()
-        previousHandlePrefix = ItemUpdate.HANDLE_PREFIX;
-        ItemUpdate.HANDLE_PREFIX = handleService.getCanonicalPrefix();
-
-        context.restoreAuthSystemState();
-
-        tempDir = Files.createTempDirectory("embargoSafetyIT");
-    }
-
-    @After
-    @Override
-    public void destroy() throws Exception {
-        ItemUpdate.HANDLE_PREFIX = previousHandlePrefix;
-        if (tempDir != null) {
-            PathUtils.deleteDirectory(tempDir);
-        }
-        super.destroy();
-    }
 
     /**
      * Verifies that a withdrawn item is left alone: withdrawal turns every READ policy into WITHDRAWN_READ,
@@ -178,7 +78,7 @@ public class EmbargoSafetyIT extends AbstractIntegrationTestWithDatabase {
         Set<Integer> idsBefore = policyIds(bitstream);
         List<String> policiesBefore = policyFingerprints(bitstream);
 
-        Run pastRun = runItemUpdate(item, dublinCore(item, Collections.singletonList("openAccess"), pastDate()));
+        Run pastRun = runItemUpdate(item, dublinCore(item, "openAccess", pastDate()));
 
         item = context.reloadEntity(item);
         bitstream = context.reloadEntity(bitstream);
@@ -202,7 +102,7 @@ public class EmbargoSafetyIT extends AbstractIntegrationTestWithDatabase {
         // The past-date run above only reaches the early return; the future-date branch is the one that
         // creates policies.
         Run futureRun =
-                runItemUpdate(item, dublinCore(item, Collections.singletonList("embargoedAccess"), futureDate()));
+                runItemUpdate(item, dublinCore(item, "embargoedAccess", futureDate()));
 
         item = context.reloadEntity(item);
         bitstream = context.reloadEntity(bitstream);
@@ -289,7 +189,7 @@ public class EmbargoSafetyIT extends AbstractIntegrationTestWithDatabase {
         Set<Integer> idsBefore = policyIds(bitstream);
         List<String> policiesBefore = policyFingerprints(bitstream);
 
-        Run pastRun = runItemUpdate(item, dublinCore(item, Collections.singletonList("openAccess"), pastDate()));
+        Run pastRun = runItemUpdate(item, dublinCore(item, "openAccess", pastDate()));
 
         bitstream = context.reloadEntity(bitstream);
 
@@ -304,7 +204,7 @@ public class EmbargoSafetyIT extends AbstractIntegrationTestWithDatabase {
         // A past date only reaches the early return; the future-date branch is where a group-restricted file
         // could gain an Anonymous policy.
         Run futureRun = runItemUpdate(item,
-                dublinCore(item, Collections.singletonList("embargoedAccess"), futureDate()));
+                dublinCore(item, "embargoedAccess", futureDate()));
 
         bitstream = context.reloadEntity(bitstream);
 
@@ -351,7 +251,7 @@ public class EmbargoSafetyIT extends AbstractIntegrationTestWithDatabase {
         assertFalse("fixture precondition: a bitstream without policies must not be readable"
                 + describe(bitstream), anonymousCanRead(bitstream));
 
-        Run run = runItemUpdate(item, dublinCore(item, Collections.singletonList("openAccess"), pastDate()));
+        Run run = runItemUpdate(item, dublinCore(item, "openAccess", pastDate()));
 
         bitstream = context.reloadEntity(bitstream);
 
@@ -386,7 +286,7 @@ public class EmbargoSafetyIT extends AbstractIntegrationTestWithDatabase {
 
         // The future-date branch is the one that writes policies, so it is where the end date would be lost.
         Run futureRun = runItemUpdate(item,
-                dublinCore(item, Collections.singletonList("embargoedAccess"), futureDate()));
+                dublinCore(item, "embargoedAccess", futureDate()));
 
         bitstream = context.reloadEntity(bitstream);
 
@@ -399,7 +299,7 @@ public class EmbargoSafetyIT extends AbstractIntegrationTestWithDatabase {
 
         // An expired end date reaches the same mutation, only with a start date that has already passed.
         Run pastRun = runItemUpdate(item,
-                dublinCore(item, Collections.singletonList("openAccess"), pastDate()));
+                dublinCore(item, "openAccess", pastDate()));
 
         bitstream = context.reloadEntity(bitstream);
 
@@ -428,7 +328,7 @@ public class EmbargoSafetyIT extends AbstractIntegrationTestWithDatabase {
         List<String> policiesBefore = policyFingerprints(bitstream);
 
         Run run = runItemUpdate(item,
-                dublinCore(item, Collections.singletonList("embargoedAccess"), futureDate()));
+                dublinCore(item, "embargoedAccess", futureDate()));
 
         bitstream = context.reloadEntity(bitstream);
 
@@ -503,7 +403,7 @@ public class EmbargoSafetyIT extends AbstractIntegrationTestWithDatabase {
         Set<Integer> idsBefore = policyIds(bitstream);
         List<String> policiesBefore = policyFingerprints(bitstream);
 
-        Run run = runItemUpdate(item, dublinCore(item, Collections.singletonList("openAccess"), pastDate()));
+        Run run = runItemUpdate(item, dublinCore(item, "openAccess", pastDate()));
 
         bitstream = context.reloadEntity(bitstream);
 
@@ -535,7 +435,7 @@ public class EmbargoSafetyIT extends AbstractIntegrationTestWithDatabase {
         };
 
         Run run = runItemUpdate(itemUpdate, item,
-                dublinCore(item, Collections.singletonList("openAccess"), pastDate()));
+                dublinCore(item, "openAccess", pastDate()));
 
         bitstream = context.reloadEntity(bitstream);
 
@@ -563,7 +463,7 @@ public class EmbargoSafetyIT extends AbstractIntegrationTestWithDatabase {
         Set<Integer> idsBefore = policyIds(bitstream);
         List<String> policiesBefore = policyFingerprints(bitstream);
 
-        Run run = runItemUpdate(item, dublinCore(item, accessRights, embargoEndDate));
+        Run run = runItemUpdate(item, dublinCoreWithAccessRights(item, accessRights, embargoEndDate));
 
         item = context.reloadEntity(item);
         bitstream = context.reloadEntity(bitstream);
@@ -588,150 +488,6 @@ public class EmbargoSafetyIT extends AbstractIntegrationTestWithDatabase {
         assertEquals("[" + scenario + "] a surviving resource policy was modified in place although nothing at"
                         + " all should have happened." + describe(bitstream),
                 policiesBefore, policyFingerprints(bitstream));
-    }
-
-    /**
-     * Runs itemupdate with both embargo fields as targets, the combination that triggers embargo
-     * synchronisation. The {@link ItemUpdate} instance is kept because it carries the failure count.
-     *
-     * @return the console output of the run and the number of embargo problems it counted
-     */
-    private Run runItemUpdate(Item item, String dublinCoreContent) throws Exception {
-        return runItemUpdate(new ItemUpdate(), item, dublinCoreContent);
-    }
-
-    /**
-     * The same run driven by a caller supplied {@link ItemUpdate}, so that a test can make one step of the
-     * synchronisation fail where a database error would.
-     */
-    private Run runItemUpdate(ItemUpdate itemUpdate, Item item, String dublinCoreContent) throws Exception {
-        Path sourceRoot = Files.createDirectory(tempDir.resolve("saf-" + System.nanoTime()));
-        // without this marker processArchive writes an undo archive next to the source directory
-        Files.createFile(sourceRoot.resolve(ItemUpdate.SUPPRESS_UNDO_FILENAME));
-
-        Path itemDir = Files.createDirectory(sourceRoot.resolve("item_000"));
-        Files.writeString(itemDir.resolve("dublin_core.xml"), dublinCoreContent, StandardCharsets.UTF_8);
-
-        DeleteMetadataAction deleteAction =
-                (DeleteMetadataAction) itemUpdate.actionMgr.getUpdateAction(DeleteMetadataAction.class);
-        deleteAction.addTargetFields(new String[] { "dc.rights.access", "dc.date.embargoend" });
-
-        AddMetadataAction addAction =
-                (AddMetadataAction) itemUpdate.actionMgr.getUpdateAction(AddMetadataAction.class);
-        addAction.addTargetFields(new String[] { "dc.rights.access", "dc.date.embargoend" });
-
-        // ItemUpdate reports to System.out only, so the operator console has to be captured to assert on it
-        ByteArrayOutputStream consoleBuffer = new ByteArrayOutputStream();
-        PrintStream originalOut = System.out;
-        PrintStream originalErr = System.err;
-        PrintStream captureStream = new PrintStream(consoleBuffer, true, StandardCharsets.UTF_8);
-
-        context.turnOffAuthorisationSystem();
-        System.setOut(captureStream);
-        System.setErr(captureStream);
-        try {
-            itemUpdate.processArchive(context, sourceRoot.toString(), null, null, true, false, true);
-        } finally {
-            captureStream.flush();
-            System.setOut(originalOut);
-            System.setErr(originalErr);
-            context.restoreAuthSystemState();
-        }
-
-        context.uncacheEntity(item);
-
-        String consoleOutput = consoleBuffer.toString(StandardCharsets.UTF_8);
-        // replay it so the failsafe -output.txt still holds the full ItemUpdate log
-        System.out.println(consoleOutput);
-        return new Run(consoleOutput, itemUpdate.embargoSyncFailures);
-    }
-
-    /**
-     * What a finished {@code itemupdate} run is judged by: its console output and its failure count.
-     */
-    private static final class Run {
-        private final String console;
-        private final int embargoSyncFailures;
-
-        private Run(String console, int embargoSyncFailures) {
-            this.console = console;
-            this.embargoSyncFailures = embargoSyncFailures;
-        }
-    }
-
-    /**
-     * Asserts the exit code the run would have produced; without it a skipped item looks synchronised.
-     */
-    private void assertExitCode(String scenario, int expectedFailures, Run run) {
-        assertEquals("[" + scenario + "] wrong number of reported embargo problems, so ItemUpdate.main() would"
-                        + " exit with " + ItemUpdate.exitStatus(0, run.embargoSyncFailures) + " instead of "
-                        + ItemUpdate.exitStatus(0, expectedFailures) + ". Console output was:"
-                        + System.lineSeparator() + run.console,
-                expectedFailures, run.embargoSyncFailures);
-    }
-
-    /**
-     * Builds a {@code dublin_core.xml} carrying zero or more {@code dc.rights.access} values.
-     *
-     * @param embargoEndDate {@code null} omits {@code dc.date.embargoend} entirely, the empty string writes a
-     *                       blank value (an empty XML element is dropped by the parser, so a single space is
-     *                       written instead)
-     */
-    private String dublinCore(Item item, List<String> accessRights, String embargoEndDate) {
-        StringBuilder sb = new StringBuilder();
-        sb.append("<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n")
-                .append("<dublin_core schema=\"dc\">\n")
-                .append("    <dcvalue element=\"identifier\" qualifier=\"uri\">")
-                .append(ItemUpdate.HANDLE_PREFIX)
-                .append(item.getHandle())
-                .append("</dcvalue>\n");
-
-        for (String accessRight : accessRights) {
-            sb.append("    <dcvalue element=\"rights\" qualifier=\"access\">")
-                    .append(accessRight)
-                    .append("</dcvalue>\n");
-        }
-
-        if (embargoEndDate != null) {
-            sb.append("    <dcvalue element=\"date\" qualifier=\"embargoend\">")
-                    .append(embargoEndDate.isEmpty() ? " " : embargoEndDate)
-                    .append("</dcvalue>\n");
-        }
-
-        sb.append("</dublin_core>");
-        return sb.toString();
-    }
-
-    private void ensureMetadataFieldExists(String element, String qualifier) throws Exception {
-        MetadataSchema dcSchema = metadataSchemaService.find(context, "dc");
-        MetadataField existingField = metadataFieldService.findByElement(context, dcSchema, element, qualifier);
-        if (existingField == null) {
-            MetadataFieldBuilder.createMetadataField(context, dcSchema, element, qualifier, null).build();
-        }
-    }
-
-    private Item createItem(String title) throws Exception {
-        context.turnOffAuthorisationSystem();
-        Item item = ItemBuilder.createItem(context, collection)
-                .withTitle(title)
-                .build();
-        context.restoreAuthSystemState();
-        return item;
-    }
-
-    /**
-     * A bitstream in the ORIGINAL bundle. It inherits the collection DEFAULT_BITSTREAM_READ and so carries one
-     * undated Anonymous READ policy, the state a freshly imported SAF item is in.
-     */
-    private Bitstream createOriginalBitstream(Item item, String name) throws Exception {
-        context.turnOffAuthorisationSystem();
-        Bitstream bitstream = BitstreamBuilder.createBitstream(context, item,
-                        new ByteArrayInputStream(("content-" + name).getBytes(StandardCharsets.UTF_8)))
-                .withName(name)
-                .withMimeType("text/plain")
-                .build();
-        context.restoreAuthSystemState();
-        return bitstream;
     }
 
     /**
@@ -780,64 +536,6 @@ public class EmbargoSafetyIT extends AbstractIntegrationTestWithDatabase {
     }
 
     /**
-     * Deletes policies one by one, because the bulk removal helpers issue an HQL delete and leave the
-     * in-memory collection stale.
-     *
-     * @param actionId action to delete, or {@link #ALL_ACTIONS} for every policy regardless of action
-     */
-    private void deletePolicies(Bitstream bitstream, int actionId) throws Exception {
-        List<ResourcePolicy> doomed = actionId == ALL_ACTIONS
-                ? allPolicies(bitstream)
-                : policiesForAction(bitstream, actionId);
-        for (ResourcePolicy policy : doomed) {
-            resourcePolicyService.delete(context, policy);
-        }
-    }
-
-    private List<ResourcePolicy> allPolicies(Bitstream bitstream) throws SQLException {
-        return new ArrayList<>(resourcePolicyService.find(context, bitstream));
-    }
-
-    private List<ResourcePolicy> policiesForAction(Bitstream bitstream, int actionId) throws SQLException {
-        return new ArrayList<>(resourcePolicyService.find(context, bitstream, actionId));
-    }
-
-    private List<ResourcePolicy> anonymousReadPolicies(Bitstream bitstream) throws SQLException {
-        return policiesForAction(bitstream, Constants.READ).stream()
-                .filter(policy -> policy.getGroup() != null && anonymousGroup.equals(policy.getGroup()))
-                .collect(Collectors.toList());
-    }
-
-    private Set<Integer> policyIds(Bitstream bitstream) throws SQLException {
-        Set<Integer> ids = new TreeSet<>();
-        for (ResourcePolicy policy : allPolicies(bitstream)) {
-            ids.add(policy.getID());
-        }
-        return ids;
-    }
-
-    private List<String> policyFingerprints(Bitstream bitstream) throws SQLException {
-        List<String> fingerprints = new ArrayList<>();
-        for (ResourcePolicy policy : allPolicies(bitstream)) {
-            fingerprints.add(fingerprint(policy));
-        }
-        Collections.sort(fingerprints);
-        return fingerprints;
-    }
-
-    private String fingerprint(ResourcePolicy policy) {
-        return String.format("id=%s action=%s group=%s eperson=%s rpType=%s rpName=%s start=%s end=%s",
-                policy.getID(),
-                Constants.actionText[policy.getAction()],
-                policy.getGroup() == null ? "<none>" : policy.getGroup().getName(),
-                policy.getEPerson() == null ? "<none>" : policy.getEPerson().getEmail(),
-                policy.getRpType(),
-                policy.getRpName(),
-                day(policy.getStartDate()),
-                day(policy.getEndDate()));
-    }
-
-    /**
      * Renders the current policies of the bitstream for failure messages, so a red build shows which policy
      * moved.
      */
@@ -854,54 +552,4 @@ public class EmbargoSafetyIT extends AbstractIntegrationTestWithDatabase {
         return sb.toString();
     }
 
-    /**
-     * Tells whether a visitor who is not logged in may read the bitstream. The authorisation state is a stack
-     * the builders push and pop, so it is drained first - otherwise every read looks allowed.
-     */
-    private boolean anonymousCanRead(Bitstream bitstream) throws SQLException {
-        EPerson savedUser = context.getCurrentUser();
-        int popped = 0;
-        while (context.ignoreAuthorization()) {
-            context.restoreAuthSystemState();
-            popped++;
-        }
-        context.setCurrentUser(null);
-        try {
-            return authorizeService.authorizeActionBoolean(context, bitstream, Constants.READ);
-        } finally {
-            context.setCurrentUser(savedUser);
-            for (int i = 0; i < popped; i++) {
-                context.turnOffAuthorisationSystem();
-            }
-        }
-    }
-
-    private String pastDate() {
-        return LocalDate.now().minusMonths(1).toString();
-    }
-
-    /**
-     * A future end date, the branch that writes resource policies; the past-date branch returns early.
-     */
-    private String futureDate() {
-        return LocalDate.now().plusYears(1).toString();
-    }
-
-    private Date startOfDayUtc(LocalDate day) {
-        return Date.from(day.atStartOfDay(ZoneOffset.UTC).toInstant());
-    }
-
-    /**
-     * Renders a date at calendar day granularity: start dates come back from the database as
-     * {@code java.sql.Date}, so comparing instants across the harness time zone would be flaky.
-     */
-    private String day(Date date) {
-        if (date == null) {
-            return "<none>";
-        }
-        if (date instanceof java.sql.Date) {
-            return ((java.sql.Date) date).toLocalDate().toString();
-        }
-        return date.toInstant().atZone(ZoneId.systemDefault()).toLocalDate().toString();
-    }
 }
